Split FlApplication into its own module.
This fixes a compile warning about an unused FL_IS_APPLICATION.
