+++ testbed/gtk/CMakeLists.txt
@@ -33,6 +33,7 @@ pkg_check_modules(GTK REQUIRED IMPORTED_TARGET gtk+-3.0)
 # Application build
 add_executable(${BINARY_NAME}
   "main.cc"
+  "fl_application.cc"
   "window_configuration.cc"
   "${FLUTTER_MANAGED_DIR}/gtk_plugin_registrant.cc"
 )
